Container installation
Schema and example changes to support container installation during image
build. As parts of what we considert to be part of the distribution move into
containers we want to support installing the containers during image build
to be able to have ready to go images that needs as few operations as possible
during firstboot.